#6 entity build - [x] 1차 완성된 엔티티 빌드
