+++ src/main/java/com/financial/apply/controller/ApplyController.java
@@ -1,0 +1,8 @@
+package com.financial.apply.controller;
+
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+public class ApplyController {
+}
