+++ src/main/java/com/financial/apply/entity/Apply.java
@@ -1,0 +1,36 @@
+package com.financial.apply.entity;
+
+import com.financial.interest.entity.Interest;
+import com.financial.member.entity.Member;
+import com.financial.product.entity.Product;
+import lombok.AccessLevel;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+import javax.persistence.*;
+import java.time.LocalDateTime;
+
+@Entity
+@Getter
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+public class Apply {
+    @Id
+    @GeneratedValue(strategy = GenerationType.AUTO)
+    @Column(name = "apply_id")
+    private Long id;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "member_id")
+    private Member member;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "product_id")
+    private Product product;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "interest_id")
+    private Interest interest;
+
+    private LocalDateTime applyTime;
+
+}
