+++ src/main/java/com/financial/apply/repository/ApplyRepository.java
@@ -1,0 +1,8 @@
+package com.financial.apply.repository;
+
+import com.financial.apply.entity.Apply;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface ApplyRepository extends JpaRepository<Apply, Long> {
+
+}
