+++ src/main/java/com/financial/apply/service/ApplyService.java
@@ -1,0 +1,7 @@
+package com.financial.apply.service;
+
+import org.springframework.stereotype.Service;
+
+@Service
+public class ApplyService {
+}
