+++ src/main/java/com/financial/bookmark/controller/BookmarkController.java
@@ -1,0 +1,8 @@
+package com.financial.bookmark.controller;
+
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+public class BookmarkController {
+
+}
