+++ src/main/java/com/financial/bookmark/service/BookmarkService.java
@@ -1,0 +1,7 @@
+package com.financial.bookmark.service;
+
+import org.springframework.stereotype.Service;
+
+@Service
+public class BookmarkService {
+}
