--- src/main/java/com/financial/entity/member/IsMember.java
@@ -1,5 +1,0 @@
-package com.financial.entity.member;
-
-public enum IsMember {
-    Y, N
-}
