--- src/main/java/com/financial/entity/member/Job.java
@@ -1,7 +1,0 @@
-package com.financial.entity.member;
-
-public enum Job {
-    OFFICE_WORKERS, PUBLIC_OFFICIAL, PROFESSION,
-    AGRICULTURAL_WORKER, BUISNESSMAN, FREELANCER,
-    HOUSEWIFE, STUDENT, SOLDIER, INOCCUPATION
-}
