--- src/main/java/com/financial/entity/member/ProductType.java
@@ -1,5 +1,0 @@
-package com.financial.entity.member;
-
-public enum ProductType {
-    DEPOSIT, SAVING, DEPOSIT_AND_SAVING
-}
