--- src/main/java/com/financial/entity/member/Survey.java
@@ -1,15 +1,0 @@
-package com.financial.entity.member;
-
-import javax.persistence.Embeddable;
-import javax.persistence.EnumType;
-import javax.persistence.Enumerated;
-
-@Embeddable
-public class Survey {
-    @Enumerated(EnumType.STRING)
-    private ProductType productType;
-    @Enumerated(EnumType.STRING)
-    private Job job;
-    private String mainBank;
-
-}
