--- src/main/java/com/financial/entity/product/BankName.java
@@ -1,5 +1,0 @@
-package com.financial.entity.product;
-
-public enum BankName {
-    SHIN_HAN, KOOK_MIN, WOO_RIE, HA_NA, NH
-}
