--- src/main/java/com/financial/entity/product/InterestRate.java
@@ -1,5 +1,0 @@
-package com.financial.entity.product;
-
-public enum InterestRate {
-    TWELVE, TWENTY_FOUR
-}
