--- src/main/java/com/financial/entity/product/JoinWay.java
@@ -1,5 +1,0 @@
-package com.financial.entity.product;
-
-public enum JoinWay {
-    PHONE, BANK, INTERNET
-}
