+++ src/main/java/com/financial/interest/controller/InterestController.java
@@ -1,0 +1,7 @@
+package com.financial.interest.controller;
+
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+public class InterestController {
+}
