+++ src/main/java/com/financial/interest/repository/InterestRepository.java
@@ -1,0 +1,7 @@
+package com.financial.interest.repository;
+
+import com.financial.interest.entity.Interest;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface InterestRepository extends JpaRepository<Interest, Long> {
+}
