+++ src/main/java/com/financial/interest/service/InterestService.java
@@ -1,0 +1,7 @@
+package com.financial.interest.service;
+
+import org.springframework.stereotype.Service;
+
+@Service
+public class InterestService {
+}
