+++ src/main/java/com/financial/keyword/controller/KeywordController.java
@@ -1,0 +1,7 @@
+package com.financial.keyword.controller;
+
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+public class KeywordController {
+}
