+++ src/main/java/com/financial/keyword/repository/KeywordRepository.java
@@ -1,0 +1,8 @@
+package com.financial.keyword.repository;
+
+import com.financial.keyword.entity.Keyword;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface KeywordRepository extends JpaRepository<Keyword, Long> {
+
+}
