+++ src/main/java/com/financial/keyword/service/KeywordService.java
@@ -1,0 +1,7 @@
+package com.financial.keyword.service;
+
+import org.springframework.stereotype.Service;
+
+@Service
+public class KeywordService {
+}
