+++ src/main/java/com/financial/member/controller/MemberController.java
@@ -1,0 +1,7 @@
+package com.financial.member.controller;
+
+import org.springframework.stereotype.Controller;
+
+@Controller
+public class MemberController {
+}
