+++ src/main/java/com/financial/member/entity/enums/IsMember.java
@@ -1,0 +1,5 @@
+package com.financial.member.entity.enums;
+
+public enum IsMember {
+    Y, N
+}
