+++ src/main/java/com/financial/member/entity/enums/ProductType.java
@@ -1,0 +1,5 @@
+package com.financial.member.entity.enums;
+
+public enum ProductType {
+    DEPOSIT, SAVING, DEPOSIT_AND_SAVING
+}
