+++ src/main/java/com/financial/member/repository/MemberRepository.java
@@ -1,0 +1,7 @@
+package com.financial.member.repository;
+
+import com.financial.member.entity.Member;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface MemberRepository extends JpaRepository<Member, Long> {
+}
