+++ src/main/java/com/financial/member/service/MemberService.java
@@ -1,0 +1,7 @@
+package com.financial.member.service;
+
+import org.springframework.stereotype.Service;
+
+@Service
+public class MemberService {
+}
