+++ src/main/java/com/financial/product/controller/ProductController.java
@@ -1,0 +1,7 @@
+package com.financial.product.controller;
+
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+public class ProductController {
+}
