+++ src/main/java/com/financial/product/entity/enums/BankName.java
@@ -1,0 +1,5 @@
+package com.financial.product.entity.enums;
+
+public enum BankName {
+    SHIN_HAN, KOOK_MIN, WOO_RIE, HA_NA
+}
