+++ src/main/java/com/financial/product/entity/enums/InterestRate.java
@@ -1,0 +1,5 @@
+package com.financial.product.entity.enums;
+
+public enum InterestRate {
+    TWELVE, TWENTY_FOUR
+}
