+++ src/main/java/com/financial/product/entity/enums/JoinWay.java
@@ -1,0 +1,5 @@
+package com.financial.product.entity.enums;
+
+public enum JoinWay {
+    PHONE, BANK, INTERNET
+}
