+++ src/main/java/com/financial/product/repository/ProductRepository.java
@@ -1,0 +1,8 @@
+package com.financial.product.repository;
+
+import com.financial.product.entity.Product;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface ProductRepository extends JpaRepository<Product, Long> {
+
+}
