+++ src/main/java/com/financial/product/service/ProductService.java
@@ -1,0 +1,7 @@
+package com.financial.product.service;
+
+import org.springframework.stereotype.Service;
+
+@Service
+public class ProductService {
+}
