+++ src/main/resources/application.properties
@@ -8,4 +8,5 @@ spring.profiles.include=profile
 # hibernate ??
 spring.jpa.properties.hibernate.format_sql = true
 spring.jpa.properties.hibernate.show_sql = true
-spring.jpa.hibernate.ddl-auto=create
+spring.jpa.properties.hibernate.default_batch_fetch_size=1000
+spring.jpa.hibernate.ddl-auto=none
